+++ src/pages/p-v4-components-demo/index.js
@@ -18,5 +18,6 @@ package('p-v4-components-demo')
 		'b-router',
 		'b-list',
 		'b-sidebar',
+		'b-window',
 		components
 	);